Fixed plotting the correlation function.
I was using keys instead of values in the dictionary. Then normalizing
was the problem. It is now normalized on the number of available sites
that can give that key in the dictionary. This do not give 1 since the
worm can 'eat itself' and move in the same space twice.